install: split into frozen/settings/side_effects_cache submodules
Moves `crates/aube/src/commands/install.rs` (5239 lines) into a
module directory with three sibling files. Pure refactor — no
behavior change.

- `install/frozen.rs`: `FrozenMode`, `GlobalFrozenFlags`,
  `GlobalVirtualStoreFlags` + their tests.
- `install/side_effects_cache.rs`: `SideEffectsCacheConfig`,
  `SideEffectsCacheEntry`, marker/hash/copy helpers.
- `install/settings.rs`: `resolve_*` settings helpers,
  `configure_resolver`, `PeerDependencyRules`, `warn_unmet_peers`.
- `install/mod.rs`: `InstallArgs`, `InstallOptions`, `run`, git-dep
  prepare, `run_dep_lifecycle_scripts`, linker helpers.

Public surface preserved via `pub use` / `pub(crate) use` in
mod.rs so external callers (`main.rs`, `ci.rs`, `deploy.rs`,
`dlx.rs`, `rebuild.rs`, etc.) compile unchanged.
